+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -164,7 +164,7 @@ class ExpoCameraView(
       shouldCreateCamera = true
     }
 
-  var ratio: CameraRatio = CameraRatio.FOUR_THREE
+  var ratio: CameraRatio = CameraRatio.ONE_ONE
     set(value) {
       field = value
       shouldCreateCamera = true
@@ -353,13 +353,21 @@ class ExpoCameraView(
       {
         val cameraProvider: ProcessCameraProvider = providerFuture.get()
 
+        previewView.scaleType = if (ratio != CameraRatio.ONE_ONE) {
+          PreviewView.ScaleType.FIT_CENTER
+        } else {
+          PreviewView.ScaleType.FILL_CENTER
+        }
+
+        val resolutionSelector = ResolutionSelector.Builder().apply {
+          if (ratio != CameraRatio.ONE_ONE) {
+            setAspectRatioStrategy(ratio.mapToStrategy())
+          }
+          setResolutionStrategy(ResolutionStrategy.HIGHEST_AVAILABLE_STRATEGY)
+        }.build()
+
         val preview = Preview.Builder()
-          .setResolutionSelector(
-            ResolutionSelector.Builder()
-              .setAspectRatioStrategy(ratio.mapToStrategy())
-              .setResolutionStrategy(ResolutionStrategy.HIGHEST_AVAILABLE_STRATEGY)
-              .build()
-          )
+          .setResolutionSelector(resolutionSelector)
           .build()
           .also {
             it.surfaceProvider = previewView.surfaceProvider
@@ -375,12 +383,7 @@ class ExpoCameraView(
               val size = Size.parseSize(pictureSize)
               setTargetResolution(size)
             } else {
-              setResolutionSelector(
-                ResolutionSelector.Builder()
-                  .setAspectRatioStrategy(ratio.mapToStrategy())
-                  .setResolutionStrategy(ResolutionStrategy.HIGHEST_AVAILABLE_STRATEGY)
-                  .build()
-              )
+              setResolutionSelector(resolutionSelector)
             }
           }
           .build()